Replace custom-made slack client with official sdk
