+++ app/infrastructure/slack-web-api/index.js
@@ -1,0 +1,5 @@
+const { WebClient } = require('@slack/web-api');
+
+const config = require('../../config');
+
+module.exports = new WebClient(config.SLACK_ACCESS_TOKEN);
