--- app/infrastructure/slack/index.js
@@ -1,5 +1,0 @@
-const chatPostMessage = require('./actions/chatPostMessage');
-
-module.exports = {
-  chatPostMessage,
-};
